pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "plato"
version = "0.1.0"
authors = [
  { name="Darleison Filho" },
]
description = "Local-first PDF to Knowledge Graph & Semantic Search Pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "networkx",
    "pydantic>=2.0",
    "pyyaml",
    "ollama",
    "matplotlib",
    "spacy",
    "pandas",
    "typer",
    "rich",
    # LlamaIndex Core & Extensions
    "llama-index-core",
    "llama-index-embeddings-huggingface",
    "llama-index-llms-ollama",
]

[project.scripts]
plato = "main:main"
